% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflows-arbitrating.r
\name{ADE_nopump}
\alias{ADE_nopump}
\title{Forecasting Arbitrated Ensemble
no pump}
\usage{
ADE_nopump(form, train, test, learner, learner.pars, ...)
}
\arguments{
\item{form}{Formula}

\item{train}{embedded time series used for training the base learners}

\item{test}{embedded time series used for testing}

\item{learner}{Character vector describing the base algorithms to be trained.
Current available implemented models are:
\describe{
   \item{MARS}{Multivariate Adaptive Regression Splines from \strong{earth} package}
   \item{PPR}{Projection Pursuit Regression from \strong{stats} package}
   \item{baggedtrees}{For a bagging ensemble for time series forecasting tasks
    (Oliveira and Torgo, 2014)}
   \item{SVM}{Support Vector Machines from \strong{kernlab} package}
   \item{GP}{Gaussian Processes from \strong{kernlab} package}
   \item{FFNN}{Feed Forward Neural Networks from \strong{nnet} package}
   \item{SVM}{Support Vector Machines from \strong{kernlab} package}
   \item{Cubist}{Rule-based Regression from \strong{Cubist} package}
   \item{RandomForest}{Random Forests from \strong{ranger} package}
   \item{GBM}{Generalized Boosted Regression from \strong{gbm} package}
   \item{GLM}{Generalized Linear Models (e.g. Ridge regression, LASSO, Elastic-Net)
   from \strong{glmnet} package}
   \item{SAE}{Stacked Autoencoder from \strong{deepnet} package}
}}

\item{learner.pars}{Named list describing the parameter of the \code{learner}. Below are
described some examples.}

\item{...}{Further parameters to pass to the function}
}
\description{
Forecasting Arbitrated Ensemble
no pump
}

