% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{ae}
\alias{ae}
\title{Computing the absolute error}
\usage{
ae(y, y_hat, ...)
}
\arguments{
\item{y}{A numeric vector representing the actual values.}

\item{y_hat}{A numeric vector representing the forecasted values.}
}
\description{
Element-wise computation of the absolute error loss function.
}
\seealso{
Other error/performance functions: \code{\link{mae}},
  \code{\link{mase}}, \code{\link{mse}},
  \code{\link{r_squared}}, \code{\link{se}}
}
\keyword{internal}
