% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflows-baselines.r
\name{AEC}
\alias{AEC}
\title{Forecaster Combination
Model according to the guidelines
of Sanchez in adaptive combination of forecasts}
\usage{
AEC(form, train, test, learner, learner.pars, ff = 0.95, ...)
}
\arguments{
\item{form}{Formula}

\item{train}{embedded time series used for training the base learners}

\item{test}{embedded time series used for testing}

\item{learner}{Character vector describing the base algorithms to be trained.}

\item{learner.pars}{Named list describing the parameter of the \code{learner}. Below are
described some examples.}

\item{ff}{forgetting factor}

\item{...}{Further parameters to pass to the function}
}
\description{
Forecaster Combination
Model according to the guidelines
of Sanchez in adaptive combination of forecasts
}
