% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-models.r
\name{arima_offline_forecast}
\alias{arima_offline_forecast}
\title{sadadadadasdadas}
\usage{
arima_offline_forecast(model, newdata)
}
\arguments{
\item{model}{model}

\item{newdata}{dn}
}
\description{
sadadadadasdadas
}
