% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-models.r
\name{arima_predict}
\alias{arima_predict}
\title{Predict function for a model created from bm_arima}
\usage{
arima_predict(model, test)
}
\arguments{
\item{model}{model created using \strong{bm_arima} function}

\item{test}{test data created using \strong{embed_timeseries} function}
}
\description{
Predict function for a model created from bm_arima
}
\keyword{internal}
