% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tse-wrapper.R
\name{auto.ensembler}
\alias{auto.ensembler}
\title{Auto Tuned Time Series Ensemble}
\usage{
auto.ensembler(timeseries, search.level = "moderate")
}
\arguments{
\item{timeseries}{A time series of class xts.}

\item{search.level}{The search for the best model can be done in three levels:
\(i\) \emph{simple}; \(ii\) moderate; and \(iii\) extensive.
Beware that an \"extensive\" search will take a lot of time to run.}
}
\value{
The output is the best performing workflow as a predictive model.
}
\description{
This function is built on top of tsensembler. It is a \"for dummies\"
implementation that performs a grid search to get the best possible
model for the given time series.
}
\examples{

\dontrun{
require(xts)
ts <- as.xts(rnorm(100L), order.by = Sys.Date() + rnorm(100L))
tse <- auto.ensembler(ts, search.level = "simple")
}

}
\seealso{
\code{\link{tseModel-class}} for details of the object class
returned. \code{\link[performanceEstimation]{performanceEstimation}} on details
of the implementation of the estimation procedure;
\code{\link[performanceEstimation]{MonteCarlo-class}} on details of the
Monte Carlo simulation method. \code{\link{TSE}} for details on the workflow
for the time series ensemble.
}
