% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble-analysis.R
\name{avgRanks}
\alias{avgRanks}
\title{avgRanks}
\usage{
avgRanks(obj, metric = "rmse")
}
\arguments{
\item{obj}{Paired Comparisons results}

\item{metric}{Metric used to rank workflows. Defaults to \strong{rmse}}
}
\value{
Data.frame with Average and Standard deviation Rank statistics
of workflows
}
\description{
This function takes as input a paired comparisons object
of workflows across several experiments and computes the
average rank and respective deviation of each workflow.
}

