% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-models.R
\name{baggedtrees}
\alias{baggedtrees}
\title{baggedtrees}
\usage{
baggedtrees(form, data, embedding.dimension, nstats, learner.pars, ...)
}
\arguments{
\item{form}{formula (e.g. target ~ .).}

\item{data}{an embedded time series in a data.frame format.}

\item{embedding.dimension}{size of the embedding vector.}

\item{nstats}{number of stats used to augment the recent time series dynamics information}

\item{learner.pars}{base learner specific set of parameters}

\item{...}{Further parameters to \code{baggedtrees}}
}
\description{
baggedtrees with different embedding dimensions and summary statistics.
This function trains a bagging ensemble of trees using rpart package implementation
of decision trees. BaggingDEp_S differs from BaggingE by using different
representations of the series. This is accomplished by using
different embedding dimensions within the ensemble. With this
property the goal is to encourage diversity into the ensemble
so to reduce the variance in the combined model.
This model is based on the work done by Oliveira and Torgo (2014)
}
\references{
Oliveira and Torgo (2014) -- Ensembles for Time Series Forecasting.
In ACML Proceedings of Asian Conference on Machine Learning. JMLR: Workshop
and Conference Proceedings, 2014.
}
