% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.r
\docType{class}
\name{base_ensemble-class}
\alias{base_ensemble-class}
\title{base_ensemble-class}
\description{
\strong{base_ensemble} is a S4 class that contains the base models
comprising the ensemble. Besides the base learning algorithms --
\code{base_models} -- base_ensemble class contains information
about other meta-data used to compute predictions for new upcoming data.
}
\section{Slots}{

\describe{
\item{\code{base_models}}{a list comprising the base models;}

\item{\code{pre_weights}}{Normalized relative weights of the base learners according to
their performance on the available data;}

\item{\code{form}}{formula;}

\item{\code{colnames}}{names of the columns of the data used to train the \strong{base_models};}

\item{\code{N}}{number of base models;}

\item{\code{model_distribution}}{base learner distribution with respect to the type of learner.
That is, the number of Decision Trees, SVMs, etc.}
}}

\keyword{internal}
