% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.r
\name{base_ensemble}
\alias{base_ensemble}
\title{base_ensemble}
\usage{
base_ensemble(base_models, pre_weights, form, colnames)
}
\arguments{
\item{base_models}{a list comprising the base models;}

\item{pre_weights}{normalized relative weights of the base learners according to
their performance on the available data;}

\item{form}{formula;}

\item{colnames}{names of the columns of the data used to train the \strong{base_models};}
}
\description{
\strong{base_ensemble} is a S4 class that contains the base models
comprising the ensemble. Besides the base learning algorithms --
\code{base_models} -- base_ensemble class contains information
about other meta-data used to compute predictions for new upcoming data.
}
