% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembling-pipes.R
\name{base_models_loss}
\alias{base_models_loss}
\title{Computing the error of base models}
\usage{
base_models_loss(Y_hat, Y, loss_function = ae)
}
\arguments{
\item{Y_hat}{predictions of the base models ("@Y_hat" slot)
from \code{\link{base_ensemble-class}} object;}

\item{Y}{true values from the time series;}

\item{loss_function}{loss function to compute. Defaults to \code{ae}, absolute
error}
}
\description{
Computing the error of base models
}
\keyword{internal}
