% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-models.r
\name{best_mvr}
\alias{best_mvr}
\title{Get best PLS/PCR model}
\usage{
best_mvr(obj, form, validation_data)
}
\arguments{
\item{obj}{PLS/PCR model object}

\item{form}{formula}

\item{validation_data}{validation data used for
predicting performances of the model by number
of principal components}
}
\description{
Get best PLS/PCR model
}
\keyword{internal}
