% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out-of-bag-pipes.r
\name{blocked_prequential}
\alias{blocked_prequential}
\title{Prequential Procedure in Blocks}
\usage{
blocked_prequential(x, nfolds, FUN, .rbind = TRUE, ...)
}
\arguments{
\item{x}{data to split into \code{nfolds} blocks;}

\item{nfolds}{number of blocks to split data into;}

\item{FUN}{to apply to train/test;}

\item{.rbind}{logical. If TRUE, the results from
FUN are \strong{rbind}ed;}

\item{...}{further parameters to FUN}
}
\description{
Prequential Procedure in Blocks
}
\seealso{
\code{\link{intraining_estimations}}
function to use as \strong{FUN} parameter.
}
\keyword{internal}
