% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-models.r
\name{bm_arima}
\alias{bm_arima}
\title{Fitting an ARIMA model
Using function \strong{auto.arima} from forecast package}
\usage{
bm_arima(y, form)
}
\arguments{
\item{y}{time series struture after applying
function \strong{embed_timeseries}}

\item{form}{formula}
}
\description{
Fitting an ARIMA model
Using function \strong{auto.arima} from forecast package
}
\keyword{internal}
