% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-models.r
\name{bm_cubist}
\alias{bm_cubist}
\title{Fit Cubist models (M5)}
\usage{
bm_cubist(form, data, lpars)
}
\arguments{
\item{form}{formula}

\item{data}{training data for building the predictive
model}

\item{lpars}{a list containing the learning parameters}
}
\description{
Learning a M5 model from training data
Parameter setting can vary in \strong{committees}
and \strong{neighbors} parameters.
}
\details{
See \code{\link[Cubist]{cubist}} for a comprehensive description.

Imports learning procedure from \strong{Cubist} package.
}
\seealso{
other learning models: \code{\link{bm_mars}};
\code{\link{bm_ppr}}; \code{\link{bm_gbm}};
\code{\link{bm_glm}}; \code{\link{bm_gaussianprocess}};
\code{\link{bm_randomforest}}; \code{\link{bm_pls_pcr}};
\code{\link{bm_ffnn}}; \code{\link{bm_svr}}

Other base learning models: \code{\link{bm_ffnn}},
  \code{\link{bm_gaussianprocess}}, \code{\link{bm_gbm}},
  \code{\link{bm_glm}}, \code{\link{bm_mars}},
  \code{\link{bm_pls_pcr}}, \code{\link{bm_ppr}},
  \code{\link{bm_randomforest}}, \code{\link{bm_svr}}
}
