% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-models.r
\name{bm_ets}
\alias{bm_ets}
\title{Fitting an Exponential Smoothing model
Using function \strong{ets} from forecast package}
\usage{
bm_ets(y, form)
}
\arguments{
\item{y}{time series struture after applying
function \strong{embed_timeseries}}

\item{form}{formula}
}
\description{
Fitting an Exponential Smoothing model
Using function \strong{ets} from forecast package
}
\keyword{internal}
