% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-models.r
\name{bm_ffnn}
\alias{bm_ffnn}
\title{Fit Feedforward Neural Networks models}
\usage{
bm_ffnn(form, data, lpars)
}
\arguments{
\item{form}{formula}

\item{data}{training data for building the predictive
model}

\item{lpars}{a list containing the learning parameters}
}
\description{
Learning a Feedforward Neural Network
model from training data.
}
\details{
Parameter setting can vary in \strong{size}, \strong{maxit},
and \strong{decay} parameters.

See \code{\link[nnet]{nnet}} for a comprehensive description.

Imports learning procedure from \strong{nnet} package.
}
\seealso{
other learning models: \code{\link{bm_mars}};
\code{\link{bm_ppr}}; \code{\link{bm_gbm}};
\code{\link{bm_glm}}; \code{\link{bm_cubist}};
\code{\link{bm_randomforest}}; \code{\link{bm_pls_pcr}};
\code{\link{bm_gaussianprocess}}; \code{\link{bm_svr}}

Other base learning models: \code{\link{bm_cubist}},
  \code{\link{bm_gaussianprocess}}, \code{\link{bm_gbm}},
  \code{\link{bm_glm}}, \code{\link{bm_mars}},
  \code{\link{bm_pls_pcr}}, \code{\link{bm_ppr}},
  \code{\link{bm_randomforest}}, \code{\link{bm_svr}}
}
