% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-models.r
\name{bm_gaussianprocess}
\alias{bm_gaussianprocess}
\title{Fit Gaussian Process models}
\usage{
bm_gaussianprocess(form, data, lpars)
}
\arguments{
\item{form}{formula}

\item{data}{training data for building the predictive
model}

\item{lpars}{a list containing the learning parameters}
}
\value{
A list containing Gaussian Processes models
}
\description{
Learning a Gaussian Process model from training
data. Parameter setting can vary in \strong{kernel}
and \strong{tolerance}. See \code{\link[kernlab]{gausspr}}
for a comprehensive description.
}
\details{
Imports learning procedure from \strong{kernlab} package.
}
\seealso{
other learning models: \code{\link{bm_mars}};
\code{\link{bm_ppr}}; \code{\link{bm_gbm}};
\code{\link{bm_glm}}; \code{\link{bm_cubist}};
\code{\link{bm_randomforest}}; \code{\link{bm_pls_pcr}};
\code{\link{bm_ffnn}}; \code{\link{bm_svr}}

Other base learning models: \code{\link{bm_cubist}},
  \code{\link{bm_ffnn}}, \code{\link{bm_gbm}},
  \code{\link{bm_glm}}, \code{\link{bm_mars}},
  \code{\link{bm_pls_pcr}}, \code{\link{bm_ppr}},
  \code{\link{bm_randomforest}}, \code{\link{bm_svr}}
}
\keyword{internal}
