% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-models.r
\name{bm_pls_pcr}
\alias{bm_pls_pcr}
\title{Fit PLS/PCR regression models}
\usage{
bm_pls_pcr(form, data, lpars)
}
\arguments{
\item{form}{formula}

\item{data}{data to train the model}

\item{lpars}{parameter setting: For this multivariate regression
model the main parameter is "method". The available options are
"kernelpls", "svdpc", "cppls", "widekernelpls", and "simpls"}
}
\description{
Learning aPartial Least Squares or
Principal Components Regression from training data
}
\details{
Parameter setting can vary in \strong{method}

See \code{\link[pls]{mvr}} for a comprehensive description.

Imports learning procedure from \strong{pls} package.
}
\seealso{
other learning models: \code{\link{bm_mars}};
\code{\link{bm_ppr}}; \code{\link{bm_gbm}};
\code{\link{bm_glm}}; \code{\link{bm_cubist}};
\code{\link{bm_randomforest}}; \code{\link{bm_gaussianprocess}};
\code{\link{bm_ffnn}}; \code{\link{bm_svr}}

Other base learning models: \code{\link{bm_cubist}},
  \code{\link{bm_ffnn}}, \code{\link{bm_gaussianprocess}},
  \code{\link{bm_gbm}}, \code{\link{bm_glm}},
  \code{\link{bm_mars}}, \code{\link{bm_ppr}},
  \code{\link{bm_randomforest}}, \code{\link{bm_svr}}
}
