% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-models.r
\name{bm_ppr}
\alias{bm_ppr}
\title{Fit Projection Pursuit Regression models}
\usage{
bm_ppr(form, data, lpars)
}
\arguments{
\item{form}{formula}

\item{data}{training data for building the predictive
model}

\item{lpars}{a list containing the learning parameters}
}
\description{
Learning a Projection Pursuit Regression
model from training data. Parameter setting
can vary in \strong{nterms} and \strong{sm.method}
parameters. See \code{\link[stats]{ppr}} for a comprehensive description.
}
\details{
Imports learning procedure from \strong{stats} package.
}
\seealso{
other learning models: \code{\link{bm_mars}};
\code{\link{bm_gaussianprocess}}; \code{\link{bm_gbm}};
\code{\link{bm_glm}}; \code{\link{bm_cubist}};
\code{\link{bm_randomforest}}; \code{\link{bm_pls_pcr}};
\code{\link{bm_ffnn}}; \code{\link{bm_svr}}

Other base learning models: \code{\link{bm_cubist}},
  \code{\link{bm_ffnn}}, \code{\link{bm_gaussianprocess}},
  \code{\link{bm_gbm}}, \code{\link{bm_glm}},
  \code{\link{bm_mars}}, \code{\link{bm_pls_pcr}},
  \code{\link{bm_randomforest}}, \code{\link{bm_svr}}
}
\concept{base learning models}
\keyword{internal}
