% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-models.r
\name{bm_randomforest}
\alias{bm_randomforest}
\title{Fit Random Forest models}
\usage{
bm_randomforest(form, data, lpars)
}
\arguments{
\item{form}{formula}

\item{data}{training data for building the predictive
model}

\item{lpars}{a list containing the learning parameters}
}
\description{
Learning a Random Forest Model
from training data. Parameter setting
can vary in \strong{num.trees} and \strong{mtry}
parameters.
}
\details{
See \code{\link[ranger]{ranger}} for a comprehensive description.

Imports learning procedure from \strong{ranger} package.
}
\seealso{
other learning models: \code{\link{bm_mars}};
\code{\link{bm_ppr}}; \code{\link{bm_gbm}};
\code{\link{bm_glm}}; \code{\link{bm_cubist}};
\code{\link{bm_gaussianprocess}}; \code{\link{bm_pls_pcr}};
\code{\link{bm_ffnn}}; \code{\link{bm_svr}}

Other base learning models: \code{\link{bm_cubist}},
  \code{\link{bm_ffnn}}, \code{\link{bm_gaussianprocess}},
  \code{\link{bm_gbm}}, \code{\link{bm_glm}},
  \code{\link{bm_mars}}, \code{\link{bm_pls_pcr}},
  \code{\link{bm_ppr}}, \code{\link{bm_svr}}
}
