% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-models.r
\name{bm_svr}
\alias{bm_svr}
\title{Fit Support Vector Regression models}
\usage{
bm_svr(form, data, lpars)
}
\arguments{
\item{form}{formula}

\item{data}{training data for building the predictive
model}

\item{lpars}{a list containing the learning parameters}
}
\description{
Learning a Support Vector Regression
model from training data.
}
\details{
Parameter setting can vary in \strong{kernel},
\strong{C}, and \strong{epsilon} parameters.

See \code{\link[kernlab]{ksvm}} for a comprehensive description.

Imports learning procedure from \strong{kernlab} package.
}
\seealso{
other learning models: \code{\link{bm_mars}};
\code{\link{bm_ppr}}; \code{\link{bm_gbm}};
\code{\link{bm_glm}}; \code{\link{bm_cubist}};
\code{\link{bm_randomforest}}; \code{\link{bm_pls_pcr}};
\code{\link{bm_ffnn}}; \code{\link{bm_gaussianprocess}}

Other base learning models: \code{\link{bm_cubist}},
  \code{\link{bm_ffnn}}, \code{\link{bm_gaussianprocess}},
  \code{\link{bm_gbm}}, \code{\link{bm_glm}},
  \code{\link{bm_mars}}, \code{\link{bm_pls_pcr}},
  \code{\link{bm_ppr}}, \code{\link{bm_randomforest}}
}
\keyword{internal}
