% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-models.r
\name{bm_tbats}
\alias{bm_tbats}
\title{Fitting an tbats model
Using function \strong{tbats} from forecast package}
\usage{
bm_tbats(y, form)
}
\arguments{
\item{y}{time series struture after applying
function \strong{embed_timeseries}}

\item{form}{formula}
}
\description{
Fitting an tbats model
Using function \strong{tbats} from forecast package
}
\keyword{internal}
