% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-models.r
\name{bm_timeseries}
\alias{bm_timeseries}
\title{Classical time series models}
\usage{
bm_timeseries(form, data, lpars)
}
\arguments{
\item{form}{form}

\item{data}{training data}

\item{lpars}{list of parameters}
}
\description{
Classical time series models
}
