% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model-xgb.r
\name{bm_xgb}
\alias{bm_xgb}
\title{Base model for XGBoost}
\usage{
bm_xgb(form, data, lpars)
}
\arguments{
\item{form}{formula}

\item{data}{Training data}

\item{lpars}{list of parameters--deprecated}
}
\description{
Base model for XGBoost
}
\keyword{internal}
