% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bootstrap}
\alias{bootstrap}
\title{Pull yourself up by your bootstraps}
\usage{
bootstrap(n)
}
\arguments{
\item{n}{Vector to choose elements from and also the size of the sample}
}
\value{
A vector of sampled id's
}
\description{
bootstrap utility function for random sampling with replacement.
}
\examples{
\dontrun{
bootstrap("a")
bootstrap(FALSE)
}
bootstrap(5)
bootstrap(5)
bootstrap(3)

}
