% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembling-pipes.R
\name{build_committee}
\alias{build_committee}
\title{Building a committee for an ADE model}
\usage{
build_committee(Y_hat, Y, lambda, omega)
}
\arguments{
\item{Y_hat}{A data.frame containing the predictions
of base models;}

\item{Y}{True values of the time interval for
which to compute the committee;}

\item{lambda}{Window size. Number of observations to take into
account to build the committee;}

\item{omega}{Committee ratio -- ratio of models to dynamically weight
across the data;}
}
\description{
Building a committee for an ADE model
}
\seealso{
Other weighting base models: \code{\link{EMASE}},
  \code{\link{get_top_models}},
  \code{\link{model_recent_performance}},
  \code{\link{model_weighting}}, \code{\link{select_best}}
}
\concept{weighting base models}
\keyword{internal}
