% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca-generalisers.r
\name{CA.ADE_hat}
\alias{CA.ADE_hat}
\title{CA generaliser using arbitrage}
\usage{
CA.ADE_hat(obj, predict_info, newdata)
}
\arguments{
\item{obj}{a \code{\link{constructive_aggregation-class}}
object class}

\item{predict_info}{predict information -- output
of \code{\link{hat_info}} function}

\item{newdata}{newdata to predict}
}
\description{
CA generaliser using arbitrage
}
\keyword{internal}
