% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca-generalisers.r
\name{CA.EWA_hat}
\alias{CA.EWA_hat}
\title{CA generaliser using exponentially weighted average}
\usage{
CA.EWA_hat(obj, predict_info)
}
\arguments{
\item{obj}{a \code{\link{constructive_aggregation-class}}
object class}

\item{predict_info}{predict information -- output
of \code{\link{hat_info}} function}
}
\description{
CA generaliser using exponentially weighted average
}
\keyword{internal}
