% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca-generalisers.r
\name{CA.Ridge_hat}
\alias{CA.Ridge_hat}
\title{CA generaliser using ridge regression}
\usage{
CA.Ridge_hat(obj, predict_info)
}
\arguments{
\item{obj}{a \code{\link{constructive_aggregation-class}}
object class}

\item{predict_info}{predict information -- output
of \code{\link{hat_info}} function}
}
\description{
CA generaliser using ridge regression
}
\keyword{internal}
