% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{combine_predictions}
\alias{combine_predictions}
\title{Combining the predictions of several models}
\usage{
combine_predictions(Y_hat, W, committee = NULL)
}
\arguments{
\item{Y_hat}{a data.frame with the predictions of the
base models;}

\item{W}{a matrix or data.frame with the weights of the
base models;}

\item{committee}{A list containing the ids of the models in the
committee.}
}
\description{
This function simply applies a weighted average,
where the predictions of the base models \strong{Y_hat}
are weighted according to their weights \strong{W}. If
a \strong{committee} is specified, only models in the committee are
weighted.
}
\keyword{internal}
