% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{combinePredictions}
\alias{combinePredictions}
\title{Combine Base Learners' Predictions}
\usage{
combinePredictions(predictions, test, topmodels = NULL, modelscores = NULL,
  modelpreweights = NULL, aggregationFUN)
}
\arguments{
\item{predictions}{list of predictions from the base models.}

\item{test}{test set to predict with.}

\item{topmodels}{list of top learners accross the test set}

\item{modelscores}{model weights accross the test set}

\item{modelpreweights}{base learners pre-weights}

\item{aggregationFUN}{Combination type function}
}
\value{
final predictions for the test set
}
\description{
This function aggregates the predictions of models according
to a given combination type.
}
\seealso{
\code{\link{predict}} for details of the predict method for the
base models. \code{\link{tseModel-class}} for details how the class of the
Time Series Ensemble model.
}
