% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca-classes-methods.r
\docType{class}
\name{committee_set-class}
\alias{committee_set-class}
\title{committee_set-class}
\description{
Class for a committee set of models. It
contains most of the information regarding
a \code{\link{constructive_aggregation-class}}
}
\section{Slots}{

\describe{
\item{\code{committee_set}}{list of subsets}

\item{\code{form}}{formula}

\item{\code{specs}}{object of class \code{\link{model_specs-class}}. Contains
the parameter setting information for training the
base models;}

\item{\code{out_of_bag}}{Out of bag observations used to compute
the subsets}

\item{\code{recent_series}}{the most recent \code{lambda} observations.}

\item{\code{lambda}}{window size used to average loss. How far to
to back in time.}

\item{\code{alpha}}{contiguity size. How long should a subset
outperform other for it to be considered in the committee}

\item{\code{aggregate_subsets}}{aggregation approach for the set of subsets.}

\item{\code{aggregate_hypos}}{final aggregation approach. How should the
combined opinions be aggregated.}
}}

\keyword{internal}
