% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{compute_predictions}
\alias{compute_predictions}
\title{Compute the predictions of base models}
\usage{
compute_predictions(M, form, data)
}
\arguments{
\item{M}{list of base models;}

\item{form}{formula;}

\item{data}{new data to predict;}
}
\description{
This function is used to predict new observations
using the predictive models comprising an ensemble.
It calls on the respective method based on the type of
model, and returns the predictions as a list.
}
\keyword{internal}
