% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{computePredictions}
\alias{computePredictions}
\title{Compute Predictions for a given base model}
\usage{
computePredictions(.Model, data, form)
}
\arguments{
\item{.Model}{Base learners comprising the ensemble.}

\item{data}{embedded time series.}

\item{form}{Formula}
}
\value{
list of predictions for each base learning model.
}
\description{
This function computes the predictions of a list of base learners
in a test set
}

