% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca-classes-methods.r
\name{constructive_aggregation_}
\alias{constructive_aggregation_}
\title{constructive_aggregation_}
\usage{
constructive_aggregation_(committee_obj, base_ensemble_obj)
}
\arguments{
\item{committee_obj}{\code{\link{committee_set-class}} object}

\item{base_ensemble_obj}{\code{\link{base_ensemble-class}} object}
}
\description{
Simpler way of creating a \code{\link{constructive_aggregation-class}}
class object, from \code{\link{committee_set-class}} and
\code{\link{base_ensemble-class}}. It assumes that
\code{\link{model_specs-class}} are consistent.
}
\keyword{internal}
