% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca-buildC.r
\name{contiguous_count}
\alias{contiguous_count}
\title{Contiguity check}
\usage{
contiguous_count(C, alpha)
}
\arguments{
\item{C}{committee set}

\item{alpha}{contiguity interval size}
}
\description{
This function checks the contiguity of
out-performance of predictive models,
and respective regions where this occurs
}
\keyword{internal}
