% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-models.R
\name{Cubist}
\alias{Cubist}
\title{Cubist}
\usage{
Cubist(form, data, learner.pars)
}
\arguments{
\item{form}{formula}

\item{data}{embedded time series}

\item{learner.pars}{list of learner parameter.}
}
\description{
Cubist is the implementation of learning procedure
of an boosted rule-based regression model based on the
implementation of \code{Cubist} package.
}

