% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.r
\docType{class}
\name{dets_hat-class}
\alias{dets_hat-class}
\title{Predictions by an DETS ensemble}
\description{
Predictions by an DETS ensemble
}
\section{Slots}{

\describe{
\item{\code{y_hat}}{combined predictions of the ensemble
\code{\link{DETS-class}}. A numeric vector;}

\item{\code{Y_hat}}{a matrix containing the predictions made by
individual models;}

\item{\code{Y_committee}}{a list describing the models selected for
predictions at each time point (according to \strong{lambda}
and \strong{omega});}

\item{\code{W}}{a matrix with the weights of the base models at each prediction
time.}
}}

\seealso{
Other ensemble predictions: \code{\link{ade_hat-class}},
  \code{\link{ade_hat}}, \code{\link{dets_hat}}
}
