% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.r
\name{dets_hat}
\alias{dets_hat}
\title{Predictions by an DETS ensemble}
\usage{
dets_hat(y_hat, Y_hat, Y_committee, W)
}
\arguments{
\item{y_hat}{combined predictions of the ensemble
\code{\link{DETS}}. A numeric vector;}

\item{Y_hat}{a matrix containing the predictions made by
individual models;}

\item{Y_committee}{a list describing the models selected for
predictions at each time point (according to \strong{lambda}
and \strong{omega});}

\item{W}{a matrix with the weights of the base models at each prediction
time.}
}
\value{
Set of results from predicting with a \code{DETS}
ensemble
}
\description{
Predictions by an DETS ensemble
}
\seealso{
Other ensemble predictions: \code{\link{ade_hat-class}},
  \code{\link{ade_hat}}, \code{\link{dets_hat-class}}
}
\concept{ensemble predictions}
\keyword{internal}
