% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.r
\name{diversity}
\alias{diversity}
\title{Diversity analysis between pairs of models}
\usage{
diversity(Y_hat, Y = NULL)
}
\arguments{
\item{Y_hat}{matrix with colwise predictions}

\item{Y}{vector of true values}
}
\description{
Methods: correlation, covariance, mutual info,
disagreement measure and chi-square
}
\details{
ref: Measuring Diversity in Regression Ensembles
author of paper: haimonti duta
}

