% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembling-pipes.R
\name{EMASE}
\alias{EMASE}
\title{Weighting Base Models by their Moving Average Squared Error}
\usage{
EMASE(loss, lambda, pre_weights)
}
\arguments{
\item{loss}{Squared error of the models at each test point;}

\item{lambda}{Number of periods to average over when computing MASE;}

\item{pre_weights}{pre-weights of the base models computed in the
train set.}
}
\value{
The weights of the models in test time.
}
\description{
This function computes the weights of the learning models
using the Moving Average Squared Error (MASE) function
This method provides a simple way to quantify the recent
performance of each base learner and adapt the combined
model accordingly.
}
\seealso{
Other weighting base models: \code{\link{build_committee}},
  \code{\link{get_top_models}},
  \code{\link{model_recent_performance}},
  \code{\link{model_weighting}}, \code{\link{select_best}}
}
\keyword{internal}
