% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-preprocess.R
\name{embedStats}
\alias{embedStats}
\title{Computing Summary Statistics on the Embedding Vectors}
\usage{
embedStats(data, stats.only = FALSE)
}
\arguments{
\item{data}{a matrix or data.frame composed by the embedding vectors}

\item{stats.only}{logical. if TRUE, it returns the stats as data.frame. Otherwise
stats are returned cbinded with the embedded series. Defaults to FALSE.}
}
\value{
A data.frame with embedding vectors binded with summary statistics
}
\description{
This function computes summary statistics on the embedding vectors
The objetive is to augment the information on the recent dynamics of
the series.
}
\examples{
\dontrun{
require(xts)
ts <- as.xts(rnorm(100L), order.by = Sys.Date() + rnorm(100L))
embedded.ts <- embedStats(embed.timeseries(ts, 20L))
}

}
\seealso{
\code{\link{embed.timeseries}} for details on the embedding
of a time series into an Euclidean space.
}

