% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classmethods-tsensembler.R
\docType{methods}
\name{ensembler,tsensembler-method}
\alias{ensembler,tsensembler-method}
\title{Generic Dynamic Ensembles for Parametric Learning Models}
\usage{
\S4method{ensembler}{tsensembler}(obj, form)
}
\arguments{
\item{obj}{object from class \strong{tsensembler}}

\item{form}{formula}
}
\description{
ensembler is a method from class \code{tsensembler}. This method
applies learns a dynamic ensemble from the dataset provided.
}
