% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classmethods-tsensembler.R
\name{ensembler}
\alias{ensembler}
\title{Generic Dynamic Ensembles for Parametric Learning Models}
\usage{
ensembler(obj, form)
}
\arguments{
\item{obj}{object from class \strong{tsensembler}}

\item{form}{formula}
}
\description{
ensembler is a method from class \code{tsensembler}. This method
applies learns a dynamic ensemble from the dataset provided.
}
