% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflows-baselines.r
\name{ERP_Timmermann}
\alias{ERP_Timmermann}
\title{Forecaster Combination
Model according to the guidelines
of Timmermann in Elusive Return Predictability}
\usage{
ERP_Timmermann(form, train, test, learner, learner.pars, ma_n = 50,
  R2_min = 0.1, ...)
}
\arguments{
\item{form}{Formula}

\item{train}{embedded time series used for training the base learners}

\item{test}{embedded time series used for testing}

\item{learner}{Character vector describing the base algorithms to be trained.}

\item{learner.pars}{Named list describing the parameter of the \code{learner}. Below are
described some examples.}

\item{ma_n}{moving avg}

\item{R2_min}{min r squared}

\item{...}{Further parameters to pass to the function}
}
\description{
Forecaster Combination
Model according to the guidelines
of Timmermann in Elusive Return Predictability
}
