% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-models.r
\name{ets_predict}
\alias{ets_predict}
\title{Predict function for a model created from bm_ets}
\usage{
ets_predict(model, test)
}
\arguments{
\item{model}{model created using \strong{bm_ets} function}

\item{test}{test data created using \strong{embed_timeseries} function}
}
\description{
Predict function for a model created from bm_ets
}
\keyword{internal}
