% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expLoss}
\alias{expLoss}
\title{Exponential Weighted Average Loss}
\usage{
expLoss(loss, N, t_i, .proportion = FALSE, windsize = NULL)
}
\arguments{
\item{loss}{loss base models in a given prediction time.}

\item{N}{number of total experts}

\item{t_i}{prediction time. This is used in order to maintain uniform
loss bounds.}

\item{.proportion}{if TRUE, scales the results.}

\item{windsize}{Window size used to compute bounds}
}
\value{
Exponential loss of each base model in a given prediction time
step.
}
\description{
This is an utility function that computes the exponential
loss of the base learners, in order to measure the regret of the combined
model to such base models. This is based on theorem 2.3 presented by
CESA-BIANCHI and LUGOSI in their book Prediction, Learning, and Games
\(2006\).
}
