% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-models.R
\name{FFNN}
\alias{FFNN}
\title{FFNN}
\usage{
FFNN(form, data, learner.pars)
}
\arguments{
\item{form}{formula}

\item{data}{embedded time series}

\item{learner.pars}{list of learner parameter.}
}
\description{
FFNN is the implementation of learning procedure
of an Feed Forward Neural Network for regression, based on the
implementation of \code{nnet} package.
}

