% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FIFO}
\alias{FIFO}
\title{First-In First Out}
\usage{
FIFO(x, .in)
}
\arguments{
\item{x}{Vector}

\item{.in}{New input value for vector x of the same class as \code{class(vector)}}
}
\value{
A new vector \code{x}
}
\description{
First-In First Out utility function inserts a new value \code{.in} into
a given sequential vector \code{x}, dropping the last value of the sequence
}
\examples{
FIFO(1:10, 11)
FIFO(LETTERS[1:10], letters[1])

}
