% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classmethods-tseModel.R
\docType{methods}
\name{forecast,tseModel-method}
\alias{forecast,tseModel-method}
\title{Forecasting with a Dynamic Ensemble for Time Series tasks.}
\usage{
\S4method{forecast}{tseModel}(object, timeseries, h)
}
\arguments{
\item{object}{\code{tseModel-class} object.}

\item{timeseries}{Time series used to train the \code{object}.}

\item{h}{Number of periods for forecasting.}
}
\value{
\code{h} forecasting points ahead of a given \code{timeseries}.
}
\description{
\code{forecast} is a method under \code{tseModel-class} domain that
predicts the next \code{h} data points given a \code{tseModel-class}
trained on a given \code{timeseries}.
}
\examples{

\dontrun{
require(xts)
ts <- as.xts(rnorm(100L), order.by = Sys.Date() + rnorm(100L))
TSEr <- tsensembler(timeseries = ts,
                    form = target ~.,
                    workflow = Workflow(wf = 'TSE',
                                        learner = c('MARS', 'PPR'),
                                        learner.pars = list(mars = list(nk = c(5, 2),
                                                                        degree= c(3, 4)),
                                                            ppr = list(nterms = c(2,3,4))),
                                        varying.embed = TRUE,
                                        varying.trainwindow = FALSE,
                                        committee.ratio = .1,
                                        aggregationFUN = "emase-scommittee",
                                        verbose = FALSE),
                    embedding.dimension = 30,
                    nReps = 2,
                    modelOutput = TRUE)
tse.model <- ensembler(TSEr)

pointforecast <- forecast(tse.model, ts, h = 4)
}

}
\seealso{
\code{\link{predict}} for a closely related method.
}

