% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out-of-bag-pipes.r
\name{ForwardValidation}
\alias{ForwardValidation}
\title{Forward Validation Procedure}
\usage{
ForwardValidation(x, nfolds, FUN, .rbind = TRUE, ...)
}
\arguments{
\item{x}{data to split into \code{nfolds}}

\item{nfolds}{number of blocks to split data into}

\item{FUN}{to apply to train/test}

\item{.rbind}{logical. rbind results from FUN?}

\item{...}{further parameters to FUN}
}
\description{
Forward Validation Procedure
}
