% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-models.R
\name{GBM}
\alias{GBM}
\title{GBM}
\usage{
GBM(form, data, learner.pars)
}
\arguments{
\item{form}{formula}

\item{data}{embedded time series}

\item{learner.pars}{list of learner parameter.}
}
\description{
GBM is the implementation of learning procedure
of an generalized boosted regression model based on the
implementation of \code{gbm} package.
}
