% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tse-pipes.r
\name{get_embedcols}
\alias{get_embedcols}
\title{get cols that are part of the embedding vector}
\usage{
get_embedcols(x)
}
\arguments{
\item{x}{embedded time series}
}
\description{
get cols that are part of the embedding vector
}
