% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tse-pipes.r
\name{get_embedsize}
\alias{get_embedsize}
\title{Get embedding dimension from an embedded time series}
\usage{
get_embedsize(x)
}
\arguments{
\item{x}{embedded time series}
}
\description{
Get embedding dimension from an embedded time series
}
