% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_target}
\alias{get_target}
\title{Get the target from a formula}
\usage{
get_target(form)
}
\arguments{
\item{form}{formula}
}
\value{
the target variable as character
}
\description{
Get the target from a formula
}
