% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembling-pipes.R
\name{get_top_models}
\alias{get_top_models}
\title{Extract top learners from their weights}
\usage{
get_top_models(scores, omega)
}
\arguments{
\item{scores}{data frame containing the weights;}

\item{omega}{committee ratio of top base learners}
}
\value{
A list containing the top base models
}
\description{
This function extracts the top learners at each test point
from a score matrix, according to the committee ratio \strong{omega}.
}
\seealso{
Other weighting base models: \code{\link{EMASE}},
  \code{\link{build_committee}},
  \code{\link{model_recent_performance}},
  \code{\link{model_weighting}}, \code{\link{select_best}}
}
\keyword{internal}
