% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_y}
\alias{get_y}
\title{Get the response values from a data matrix}
\usage{
get_y(data, form)
}
\arguments{
\item{data}{data set with the response values;}

\item{form}{formula}
}
\description{
Given a formula and a data set, \code{get_y} function retrieves
the response values.
}
