% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembler.R
\name{getToplearners}
\alias{getToplearners}
\title{Extract Top Learners from the Weights of Base Models}
\usage{
getToplearners(scores, committee.ratio)
}
\arguments{
\item{scores}{data frame containing the scores}

\item{committee.ratio}{ratio of base learners to keep}
}
\value{
Top learners in test time
}
\description{
This function extracts the top learners at each test point
from a score matrix, according to the committee ratio.
}

