% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classmethods-ComparisonResults.R
\docType{methods}
\name{getTSE,ComparisonResults-method}
\alias{getTSE,ComparisonResults-method}
\title{Getting the best TSE model from a given Monte Carlo iteration}
\usage{
\S4method{getTSE}{ComparisonResults}(object, .iter = 1L)
}
\arguments{
\item{object}{\code{ComparisonResults} object}

\item{.iter}{Monte Carlo repetition}
}
\value{
A tseModel-class object
}
\description{
This utility function handles a \code{ComparisonResults} object to
extract the best performing learning ensemble model in a given \code{.iter}
Monte Carlo repetition.
}

