% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-models.R
\name{GLM}
\alias{GLM}
\title{GLM}
\usage{
GLM(form, data, learner.pars)
}
\arguments{
\item{form}{formula}

\item{data}{embedded time series}

\item{learner.pars}{list of learner parameter.}
}
\description{
learningGLM is the implementation of learning procedure
of an generalized linear model (e.g. LASSO, Ridge, Elastic-net) based
on the implementation of \code{glmnet} package.
}
