% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-models.R
\name{GP}
\alias{GP}
\title{GP}
\usage{
GP(form, data, learner.pars)
}
\arguments{
\item{form}{formula}

\item{data}{embedded time series}

\item{learner.pars}{list of learner parameter.}
}
\description{
learningGP is the implementation of learning procedure
of a Gaussian Processes model based on the implementation
of \code{kernlab} package.
}

