% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classmethods-tsesearch.R
\docType{methods}
\name{gs.ensembler,tsesearch-method}
\alias{gs.ensembler,tsesearch-method}
\title{Learning a Dynamic Ensemble for Time Series Forecasting}
\usage{
\S4method{gs.ensembler}{tsesearch}(obj)
}
\arguments{
\item{obj}{An object of class \code{tsensembler}.}
}
\value{
If the \emph{modelOutput} parameter from \code{tsensembler-class} object is set to TRUE,
the function will return a predictive model ensemble of class \code{tseModel}. Otherwise,
performance metrics are return.
}
\description{
This is the implementation of the learning procedure of the ensemble method.
The application is run on an object of class \code{tsensembler}, which provide
the options about how the ensemble should be construct such as: the embedding
dimension; the number of Monte Carlo repetitions, etc.
}
\details{
This method is built on top of \code{\link[performanceEstimation]{performanceEstimation}}.
It provides a framework for estimating the performance of many workflows.
}
\examples{

\dontrun{
require(xts)
ts <- as.xts(rnorm(100L), order.by = Sys.Date() + rnorm(100L))
TSEr <- tsensembler(timeseries = ts,
                    form = target ~.,
                    workflow = Workflow(wf = 'TSE',
                                        learner = c('MARS', 'PPR'),
                                        learner.pars = list(mars = list(nk = c(5, 2),
                                                                        degree= c(3, 4)),
                                                            ppr = list(nterms = c(2,3,4))),
                                        varying.embed = TRUE,
                                        varying.trainwindow = FALSE,
                                        committee.ratio = .1,
                                        aggregationFUN = "emase-scommittee",
                                        verbose = FALSE),
                    embedding.dimension = 30,
                    nReps = 2,
                    modelOutput = TRUE)
tse.model <- ensembler(TSEr)
}

}
\seealso{
\code{\link{tsensembler-class}} for details about the \strong{obj} and
\code{\link{tseModel-class}} for details on the generated model.
\code{\link[performanceEstimation]{MonteCarlo-class}} comprises an explanation
about the Monte Carlo simulation procedure. \code{\link{TSE}} for details on the workflow
for the time series ensemble.
}

