% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca-integration.r
\name{hat_info}
\alias{hat_info}
\title{Get predict data for generalising}
\usage{
hat_info(object, newdata)
}
\arguments{
\item{object}{a \code{\link{constructive_aggregation-class}} object}

\item{newdata}{new data for prediction}
}
\description{
Get predict data for generalising
}
\keyword{internal}
