% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out-of-bag-pipes.r
\name{holdout}
\alias{holdout}
\title{holdout Procedure for oob}
\usage{
holdout(x, estimation.ratio, FUN, ...)
}
\arguments{
\item{x}{data}

\item{estimation.ratio}{ratio of obsrevation to train}

\item{FUN}{to apply to train/test}

\item{...}{further parameters to FUN}
}
\description{
holdout Procedure for oob
}
