% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out-of-bag-pipes.r
\name{intraining_predictions}
\alias{intraining_predictions}
\title{Out-of-bag predictions}
\usage{
intraining_predictions(train, test, form, specs)
}
\arguments{
\item{train}{train set from the training set;}

\item{test}{test set from the training set;}

\item{form}{formula;}

\item{specs}{object of class \code{\link{model_specs-class}}. Contains
the specifications of the base models.}
}
\description{
A pipeline for retrieving out-of-bag predictions
from the base models
}
\seealso{
Other out-of-bag functions: \code{\link{intraining_estimations}}
}
\concept{out-of-bag functions}
\keyword{internal}
