% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{l1apply}
\alias{l1apply}
\title{Applying lapply on the rows}
\usage{
l1apply(obj, FUN, ...)
}
\arguments{
\item{obj}{a data.frame object to apply the function.}

\item{FUN}{function to apply to each row of \code{obj}}

\item{...}{Further parameters to \code{lapply}}
}
\description{
Wrapper function used to compute lapply on the rows
of a data.frame
}
\keyword{internal}
