% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-learn.r
\name{lambda.delegation}
\alias{lambda.delegation}
\title{lambda delegation function}
\usage{
lambda.delegation(Y_hat, Y, test, lwindow, lambda)
}
\arguments{
\item{Y_hat}{predictions made by the base models}

\item{Y}{true values}

\item{test}{test set}

\item{lwindow}{recent window size to compute \strong{lambda threshold}.}

\item{lambda}{accuracy threshold}
}
\description{
lambda delegation function
}
