% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-learn.r
\name{lambda.prediction}
\alias{lambda.prediction}
\title{lambda prediction of arbitrated forecasters}
\usage{
lambda.prediction(test, E_hat, Y_hat, Y, lwindow, lambda)
}
\arguments{
\item{test}{test set}

\item{E_hat}{oM predictions of M ae}

\item{Y_hat}{predictions made by the base models}

\item{Y}{true values}

\item{lwindow}{recent window size to compute \strong{lambda threshold}.}

\item{lambda}{accuracy threshold}
}
\description{
the idea is to track the performance of the base models
those who have an accuracy - which is defined as REF - below a
lambda threshold are left out of the prediction.
}
