% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembler.R
\name{LearnerAnalysis}
\alias{LearnerAnalysis}
\title{Dynamic Analysis of Performance of Base Learners}
\usage{
LearnerAnalysis(predictions, ma.N, committee.ratio = 0.1, preweights,
  aggregationFUN, Mstar = NULL)
}
\arguments{
\item{predictions}{A list of data.frames comprising the predictions
of the base models. Each \code{data.frame} contains
information on the performance of a model: One column with the true values
and a second column with the respective predictions}

\item{ma.N}{Number of periods to average over when computing MASE.}

\item{committee.ratio}{Proportion of models (top performing ones) used in
each prediction time.}

\item{preweights}{The initial weights of the models, computed in
the available data.}

\item{aggregationFUN}{Manner of aggregation of learners. Basic form is to
avegare the prediction of all learners.}

\item{Mstar}{todo}
}
\value{
The weights of each learning model (fScores). If applicable, it also returns
the top performing models (TOPmodels)
}
\description{
This function dynamically weights learning models according to
their performance.
}

