% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembler.R
\name{learnM}
\alias{learnM}
\title{Wrapper for learning the base learning models M
It encapsulates the functions \code{Learntse} and \code{tseModel}}
\usage{
learnM(form, data, learner, learner.pars, embedding.dimension)
}
\arguments{
\item{form}{formula}

\item{data}{training data}

\item{learner}{learner info}

\item{learner.pars}{parameter setting for the models}

\item{embedding.dimension}{k size}
}
\description{
Wrapper for learning the base learning models M
It encapsulates the functions \code{Learntse} and \code{tseModel}
}

