% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembler.R
\name{Learntse}
\alias{Learntse}
\title{Learning the Base Learners that comprise the Ensemble}
\usage{
Learntse(train, form, learner, learner.pars, Ksplits = NULL, Rsplits = NULL,
  nstats, embedding.dimension, verbose)
}
\arguments{
\item{train}{Training set}

\item{form}{Formula}

\item{learner}{Character describing the learning models}

\item{learner.pars}{List of parameters regarding the learning models.}

\item{Ksplits}{Column-wise (predictors) splits to be made}

\item{Rsplits}{Row-wise (training window) splits to be made}

\item{nstats}{Number of descriptive statistics used.}

\item{embedding.dimension}{todo}

\item{verbose}{Logical. If TRUE, some information about the current status of the
training procedure will be printed into the console.}
}
\value{
A series of predictive models (\code{learningModels}), the weights of
the models computed in the training data (\code{preweights}) and the
maximum expected loss of all models in the data (\code{Mstar})
}
\description{
Learntse function applies the learning algorithms
inside the \code{TSE} workflow. It receives as input a training dataset
and the meta-data about the learning algorithms. The output is a series of
predictive models, the relative performance of these models
in the training data and the maximum expected loss incurred by
the learning models in the data.
}
\seealso{
\code{\link{tseModel-class}} on detail for building an ensemble
with the results from \emph{Learntse}. \code{\link{predict}} for details on
the predict method for \emph{tseModel} class objects.
}

