% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classmethods-hats.R
\name{loss_M}
\alias{loss_M}
\title{The individual models compute point-wise loss}
\usage{
loss_M(object, prop = TRUE, lossFUN = se)
}
\arguments{
\item{object}{object of class \code{tse_hat}, the predictions
made by a \code{tseModel} object.}

\item{prop}{Logical for proportionalization. If \strong{TRUE}
results are return as \code{data.frame}. Otherwise returns
as \code{list}. Defaults to TRUE}

\item{lossFUN}{loss function to compute across base learners}
}
\description{
The individual models compute point-wise loss
}

