% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-modeling.R
\name{loss_meta_learn}
\alias{loss_meta_learn}
\title{Training an arbiter}
\usage{
loss_meta_learn(form, data, meta_model)
}
\arguments{
\item{form}{form}

\item{data}{data}

\item{meta_model}{learning algorithm -- either a "randomforest",
a "lasso", or a "gaussianprocess".}
}
\description{
Training an arbiter
}
\keyword{internal}
