% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{mae}
\alias{mae}
\title{mae}
\usage{
mae(y, y_hat)
}
\arguments{
\item{y}{A numeric vector representing the actual values.}

\item{y_hat}{A numeric vector representing the forecasted values.}
}
\description{
Mean Absolute Error loss function.
}
