% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-models.R
\name{MARS}
\alias{MARS}
\title{MARS}
\usage{
MARS(form, data, learner.pars)
}
\arguments{
\item{form}{formula}

\item{data}{embedded time series}

\item{learner.pars}{list of learner parameter.}
}
\description{
MARS is the implementation of learning procedure
of an Multivariate Adaptive Regression Splines based on the
implementation of \code{earth} package.
}

