% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{mase}
\alias{mase}
\title{Mean absolute Scaled Error}
\usage{
mase(y, y_hat)
}
\arguments{
\item{y}{A numeric vector representing the actual values.}

\item{y_hat}{A numeric vector representing the forecasted values.}
}
\description{
Mean absolute scaled error using the previous value as
baseline (i.e. a random walk)
}
