% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembler.R
\name{meanae.delegation}
\alias{meanae.delegation}
\title{from Y_hat and Y calculate mean ae and committee}
\usage{
meanae.delegation(Y_hat, Y, lambda, committee.ratio = 0.5)
}
\arguments{
\item{Y_hat}{predictions}

\item{Y}{true values}

\item{lambda}{number of observations}

\item{committee.ratio}{committee ratio}
}
\description{
from Y_hat and Y calculate mean ae and committee
}

