% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflows-baselines.r
\name{meanhat}
\alias{meanhat}
\title{Dummy mean workflow}
\usage{
meanhat(form, train, test, ...)
}
\arguments{
\item{form}{Formula -- e.g. \code{target ~.}}

\item{train}{embedded time series for training}

\item{test}{embedded time series for testing}

\item{...}{Further parameters to \code{meanhat} function}
}
\description{
Dummy mean workflow
}
