% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca-integration.r
\name{merging_in_experts}
\alias{merging_in_experts}
\title{Merge models in each committee}
\usage{
merging_in_experts(C, Y_hat, W, FUN)
}
\arguments{
\item{C}{list of ids representing
the models in each subset}

\item{Y_hat}{data frame with the prediction of each model
across data}

\item{W}{weights of each model according to recent performance}

\item{FUN}{function used to combine the models in each
subset. Currently implemented options are: window_loss,
simple, and blast.}
}
\description{
Merge models in each committee
}
\keyword{internal}
