% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-modeling.R
\name{meta_cubist}
\alias{meta_cubist}
\title{Training a RBR arbiter}
\usage{
meta_cubist(form, data)
}
\arguments{
\item{form}{formula}

\item{data}{data}
}
\description{
Training a RBR arbiter
}
\keyword{internal}
