% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-modeling.R
\name{meta_cubist_predict}
\alias{meta_cubist_predict}
\title{Arbiter predictions via Cubist}
\usage{
meta_cubist_predict(meta_model, newdata)
}
\arguments{
\item{meta_model}{arbiter -- a ranger object}

\item{newdata}{new data to predict}
}
\description{
Arbiter predictions via Cubist
}
\keyword{internal}
