% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-modeling.R
\name{meta_ffnn}
\alias{meta_ffnn}
\title{Training a Gaussian prosadacess arbiter}
\usage{
meta_ffnn(form, data)
}
\arguments{
\item{form}{form}

\item{data}{data}
}
\description{
Training a Gaussian prosadacess arbiter
}
\keyword{internal}
