% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-modeling.R
\name{meta_gp}
\alias{meta_gp}
\title{Training a Gaussian process arbiter}
\usage{
meta_gp(form, data)
}
\arguments{
\item{form}{form}

\item{data}{data}
}
\description{
Training a Gaussian process arbiter
}
