% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-modeling.R
\name{meta_gp_predict}
\alias{meta_gp_predict}
\title{Arbiter predictions via linear model}
\usage{
meta_gp_predict(model, newdata)
}
\arguments{
\item{model}{arbiter -- a Gaussian process model}

\item{newdata}{new data to predict loss}
}
\description{
Arbiter predictions via linear model
}
