% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-modeling.R
\name{meta_lasso}
\alias{meta_lasso}
\title{Training a LASSO arbiter}
\usage{
meta_lasso(form, data)
}
\arguments{
\item{form}{form}

\item{data}{data}
}
\description{
Training a LASSO arbiter
}
\keyword{internal}
