% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-modeling.R
\name{meta_lasso_predict}
\alias{meta_lasso_predict}
\title{Arbiter predictions via linear model}
\usage{
meta_lasso_predict(meta_model, newdata)
}
\arguments{
\item{meta_model}{arbiter -- a glmnet object}

\item{newdata}{new data to predict}
}
\description{
Arbiter predictions via linear model
}
\keyword{internal}
