% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-modeling.R
\name{meta_pls}
\alias{meta_pls}
\title{Training a pls process arbiter}
\usage{
meta_pls(form, data)
}
\arguments{
\item{form}{form}

\item{data}{data}
}
\description{
Training a pls process arbiter
}
\keyword{internal}
