% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-modeling.R
\name{meta_pls_predict}
\alias{meta_pls_predict}
\title{Arbiter predictions via pls model}
\usage{
meta_pls_predict(model, newdata)
}
\arguments{
\item{model}{arbiter -- a Gaussian process model}

\item{newdata}{new data to predict loss}
}
\description{
Arbiter predictions via pls model
}
\keyword{internal}
