% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-modeling.R
\name{meta_ppr}
\alias{meta_ppr}
\title{Training a meta_mars process arbiter}
\usage{
meta_ppr(form, data)
}
\arguments{
\item{form}{form}

\item{data}{data}
}
\description{
Training a meta_mars process arbiter
}
\keyword{internal}
