% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-modeling.R
\name{meta_ppr_predict}
\alias{meta_ppr_predict}
\title{Arbiter predictions via ppr model}
\usage{
meta_ppr_predict(model, newdata)
}
\arguments{
\item{model}{arbiter -- a Gaussian process model}

\item{newdata}{new data to predict loss}
}
\description{
Arbiter predictions via ppr model
}
\keyword{internal}
