% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-modeling.R
\name{meta_predict}
\alias{meta_predict}
\title{Predicting loss using arbiter}
\usage{
meta_predict(model, newdata, meta_model)
}
\arguments{
\item{model}{arbiter model}

\item{newdata}{new data to predict loss}

\item{meta_model}{learning algorithm -- either a "randomforest",
a "lasso", or a "gaussianprocess".}
}
\description{
Predicting loss using arbiter
}
