% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-modeling.R
\name{meta_rf_predict}
\alias{meta_rf_predict}
\title{Arbiter predictions via ranger}
\usage{
meta_rf_predict(meta_model, newdata)
}
\arguments{
\item{meta_model}{arbiter -- a ranger object}

\item{newdata}{new data to predict}
}
\description{
Arbiter predictions via ranger
}
