% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-modeling.R
\name{meta_svr}
\alias{meta_svr}
\title{Training a Gaussian process arbiter}
\usage{
meta_svr(form, data)
}
\arguments{
\item{form}{form}

\item{data}{data}
}
\description{
Training a Gaussian process arbiter
}
\keyword{internal}
