% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-modeling.R
\name{meta_xgb}
\alias{meta_xgb}
\title{Training a xgb arbiter}
\usage{
meta_xgb(form, data)
}
\arguments{
\item{form}{formula}

\item{data}{data}
}
\description{
Training a xgb arbiter
}
\keyword{internal}
