% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-modeling.R
\name{meta_xgb_predict}
\alias{meta_xgb_predict}
\title{Arbiter predictions via xgb}
\usage{
meta_xgb_predict(meta_model, newdata)
}
\arguments{
\item{meta_model}{arbiter -- a ranger object}

\item{newdata}{new data to predict}
}
\description{
Arbiter predictions via xgb
}
