% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.r
\name{model_accuracy}
\alias{model_accuracy}
\title{Relative accuracy measure for regression}
\usage{
model_accuracy(Y_hat, Y)
}
\arguments{
\item{Y_hat}{predictions as df}

\item{Y}{true values}
}
\value{
Logical \code{data.frame} with the accuracy of each model
}
\description{
This function computes the accuracy of regression models
according to the predictions of all models and the actual value.
}

