% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembler.R
\name{model_weighting}
\alias{model_weighting}
\title{Model Weighting with Erfc}
\usage{
model_weighting(x, trans = "softmax", ...)
}
\arguments{
\item{x}{Loss metric value}

\item{trans}{Character value describing the transformation type. 
The available options are \strong{softmax}, \strong{linear} and
\strong{erfc}. The softmax and erfc provide a non-linear transformation
where the weights decay exponentially as the relative loss of a given model 
increases (with respect to all available models). The linear transformation
is a simple normalization of values.}

\item{...}{Further arguments to \code{normalizeMaxMin} and
\code{proportion} functions (e.g. na.rm = TRUE)}
}
\value{
Weights of models
}
\description{
This is an utility function that takes the raw error of models and scales
them into a 0-1 range.
}
\details{
One can use a simple linear transformation (\code{linear}) or more
non-linear ones such as the \code{erfc} function or the \code{softmax}. 
These tranformations culminate into the final weights of the models.
}
