% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembler.R
\name{modelWeighting}
\alias{modelWeighting}
\title{Model Weighting with Erfc}
\usage{
modelWeighting(err, erfc.curve = TRUE, err.ratio = TRUE, ...)
}
\arguments{
\item{err}{Model performance metric (SE)}

\item{erfc.curve}{Logical: transform erro with erfc curve. Defaults to TRUE.}

\item{err.ratio}{Logical: Compute error proportion. Defaults to TRUE.}

\item{...}{Further arguments to \code{normalizeMaxMin} and
\code{proportion} functions (e.g. na.rm = TRUE)}
}
\value{
Weights of models
}
\description{
This is an utility function that takes the raw error of models and scales
them into a 0-1 range with the \code{erfc} function. These tranformations
culminate into the final weights of the models.
}
