% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{mse}
\alias{mse}
\title{mse}
\usage{
mse(y, y_hat)
}
\arguments{
\item{y}{A numeric vector representing the actual values.}

\item{y_hat}{A numeric vector representing the forecasted values.}
}
\description{
Utility function to compute mean squared error (MSE)
}
