% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble-analysis.R
\name{Multi.pairedComparisons}
\alias{Multi.pairedComparisons}
\title{pairedComparisons on Multiple Time Series \code{ComparisonResults} objects}
\usage{
Multi.pairedComparisons(estimationResults, baseline)
}
\arguments{
\item{estimationResults}{A list of \code{ComparisonResults} objects.}

\item{baseline}{Character name of workflow to compared methods to.}
}
\value{
Paired Comparison results accross the several predictive tasks.
}
\description{
This function applies the \code{\link[performanceEstimation]{pairedComparisons}} method
to a list of \code{ComparisonResults} objects. This can be used to analyze differences
in performance of two or more models accross several predictive tasks.
}
\seealso{
Check \code{\link[performanceEstimation]{pairedComparisons}} for a comprehensive
explanation of the core method.
}

