% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble-analysis.R
\name{Multi.WinLoss}
\alias{Multi.WinLoss}
\title{Statistical Wins and Losses on Multiple Time Series}
\usage{
Multi.WinLoss(ResultsList, test = "t.test", pretty = TRUE, ...)
}
\arguments{
\item{ResultsList}{Experiment results, process by \code{Multi.pairedComparisons} function}

\item{test}{Statistical significance test to perform. Defaults to \code{WilcoxonSignedRank.test}.
The other option is the \code{t.test}}

\item{pretty}{Logical: If TRUE, results are rendered in a more visually appealling way.}

\item{...}{Further arguments to pass to \code{WinLoss} function.}
}
\value{
a data.frame with significance results.
}
\description{
Utility function built on top of \code{WinLoss} function. This function applies and combines
the results of \code{WinLoss} accross several experiments.
}
\examples{

\dontrun{

pairedComps <- Multi.pairedComparisons(list_of_ComparisonResults, baseline)
wl <- Multi.WinLoss(pairedComps, "WilcoxonSignedRank.test")

}

}
\seealso{
\code{\link{Multi.pairedComparisons}}
}

