% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-methods.r
\docType{methods}
\name{multi_step_predict,base_ensemble-method}
\alias{multi_step_predict,base_ensemble-method}
\title{multi step prediction}
\usage{
\S4method{multi_step_predict}{base_ensemble}(object, newdata, h)
}
\arguments{
\item{object}{sasd}

\item{newdata}{nd}

\item{h}{horizon}
}
\description{
multi step prediction
}
