% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-methods.r
\name{multi_step_predict}
\alias{multi_step_predict}
\title{multi step prediction s}
\usage{
multi_step_predict(object, newdata, h)
}
\arguments{
\item{object}{sasds}

\item{newdata}{nds}

\item{h}{horizons}
}
\description{
multi step prediction s
}
