% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize}
\alias{normalize}
\title{Scale a numeric vector using max-min}
\usage{
normalize(x, ...)
}
\arguments{
\item{x}{a numeric vector.}

\item{...}{Further arguments to min and max function
(e.g. na.rm = TRUE)}
}
\value{
a linearly normalized vector
}
\description{
Utility function used to linearly normalize a numeric vector
}
\examples{
normalize(rnorm(4L))
normalize(1:10)
normalize(c(1,2,NA,4), na.rm = TRUE)

}
\keyword{internal}
