% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalizeMaxMin}
\alias{normalizeMaxMin}
\title{normalizeMaxMin}
\usage{
normalizeMaxMin(score, ...)
}
\arguments{
\item{score}{a numeric vector.}

\item{...}{Further arguments to min and max function (e.g. na.rm = TRUE)}
}
\value{
a normalized vector
}
\description{
Utility function used to normalize a numeric vector using Max-Min
normalization algorithm.
}
\examples{
normalizeMaxMin(rnorm(4L))
normalizeMaxMin(1:10)
normalizeMaxMin(c(1,2,NA,4), na.rm = TRUE)

}

