% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out-of-bag-pipes.r
\name{OOB.fun.hat}
\alias{OOB.fun.hat}
\title{oob function for predictions fun
it return out of bag predictions}
\usage{
OOB.fun.hat(train, test, form, learner, learner.pars, embedding.dimension)
}
\arguments{
\item{train}{train set from the training set}

\item{test}{test set from the training set}

\item{form}{formula}

\item{learner}{base learners}

\item{learner.pars}{pars}

\item{embedding.dimension}{k size}
}
\description{
oob function for predictions fun
it return out of bag predictions
}

