% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out-of-bag-pipes.r
\name{OOB.fun}
\alias{OOB.fun}
\title{out-of-bag function}
\usage{
OOB.fun(train, test, form, learner, learner.pars, embedding.dimension)
}
\arguments{
\item{train}{train set from the training set}

\item{test}{test set from the training set}

\item{form}{formula}

\item{learner}{base learners}

\item{learner.pars}{pars}

\item{embedding.dimension}{k size}
}
\value{
M.ae containing the ae of M and oob.train, which contains
the oob samples
}
\description{
use training data to get oob samples to train the meta-learners
}
