% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{perf_diff}
\alias{perf_diff}
\title{Log Percentual Difference}
\usage{
perf_diff(x, y)
}
\arguments{
\item{x}{baseline estimator. Numeric vector}

\item{y}{Estimator to compare with the baseline. Numeric Vector}
}
\value{
Logarithmic percentual difference
between the two estimators
}
\description{
This function computes the log percentual difference
between two numeric vector estimators
}

