% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-pointlearn.r
\name{point.metalearn.gausspr}
\alias{point.metalearn.gausspr}
\title{point metalearn with GP}
\usage{
point.metalearn.gausspr(metadata, timeseq, steps = 10)
}
\arguments{
\item{metadata}{metadata from \code{setup.metadata}}

\item{timeseq}{time sequence of test set}

\item{steps}{metamodels are retrained every \code{steps}}
}
\description{
point metalearn with GP
}

