% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointlearn.r
\name{point.metalearn.rpart}
\alias{point.metalearn.rpart}
\title{pointwise metalearn with rpart}
\usage{
point.metalearn.rpart(metadata, timeseq, steps = 10)
}
\arguments{
\item{metadata}{metadata from \code{setup.metadata}}

\item{timeseq}{time sequence of test set}

\item{steps}{metamodels are retrained every \code{steps}}
}
\description{
pointwise metalearn with rpart
}
