% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointlearn.r
\name{pointwise_M_hat_seq}
\alias{pointwise_M_hat_seq}
\title{Pointwise train of models}
\usage{
pointwise_M_hat_seq(train, test, form, learner, learner.pars)
}
\arguments{
\item{train}{train set}

\item{test}{test set}

\item{form}{formula}

\item{learner}{learning models}

\item{learner.pars}{learning model parameters}
}
\description{
Pointwise train of models
}
