% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-models.R
\name{PPR}
\alias{PPR}
\title{PPR}
\usage{
PPR(form, data, learner.pars)
}
\arguments{
\item{form}{formula}

\item{data}{embedded time series}

\item{learner.pars}{list of learner parameter.}
}
\description{
learningPPR is the implementation of learning procedure
of a Projection Pursuit Regression model based on the
implementation of \code{stats} package.
}

