% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca-classes-methods.r
\docType{methods}
\name{predict,constructive_aggregation-method}
\alias{predict,constructive_aggregation-method}
\title{predict method for constructive aggregation}
\usage{
\S4method{predict}{constructive_aggregation}(object, newdata)
}
\arguments{
\item{object}{a \code{\link{constructive_aggregation-class}} object}

\item{newdata}{new observations for prediction}
}
\description{
Method for predicting new observations using
a \code{\link{constructive_aggregation-class}} object
}
\keyword{internal}
