% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classmethods-tseModel.R
\docType{methods}
\name{predict,tseModel-method}
\alias{predict,tseModel-method}
\title{Predicting on new data with a tseModel class object}
\usage{
\S4method{predict}{tseModel}(object, newdata)
}
\arguments{
\item{object}{A \strong{tseModel-class} object.}

\item{newdata}{An embedded time series with new data points to predict.}
}
\value{
predict method for \code{tseModel} class returns four objects:
\describe{
   \item{base.predictions}{A list containing the predictions of the base learners}
   \item{predictions}{The final predictions}
   \item{baseScore}{The weights of the base learners along the new data}
   \item{topBaseLearners}{The top \code{committee.ratio} learners along the new data}
}
}
\description{
This is a \code{predict} method for predicting new data points using a
\code{tseModel} class object - refering to an ensemble for time series
forecasting.
}
\examples{

\dontrun{
require(xts)
ts <- as.xts(rnorm(100L), order.by = Sys.Date() + rnorm(100L))
TSEr <- tsensembler(timeseries = ts,
                    form = target ~.,
                    workflow = Workflow(wf = 'TSE',
                                        learner = c('MARS', 'PPR'),
                                        learner.pars = list(mars = list(nk = c(5, 2),
                                                                        degree= c(3, 4)),
                                                            ppr = list(nterms = c(2,3,4))),
                                        varying.embed = TRUE,
                                        varying.trainwindow = FALSE,
                                        committee.ratio = .1,
                                        aggregationFUN = "emase-scommittee",
                                        verbose = FALSE),
                    embedding.dimension = 30,
                    nReps = 2,
                    modelOutput = TRUE)
tse.model <- ensembler(TSEr)
newdata <- embedStats(embed.timeseries(ts, embedding.dimension))

preds <- predict(tse.model, newdata)
}

}
\seealso{
\code{\link{tseModel-class}} for details about the parameter setting
of the \code{object}, \code{\link{computePredictions}} for details on
the computation of predictions of the base learning models;
\code{\link{LearnerAnalysis}} for details on the weighting of models along the
test observations; \code{\link{combinePredictions}} on details on how the
base learners' predictions are combined.
}
