% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-models.r
\name{predict_pls_pcr}
\alias{predict_pls_pcr}
\title{predict method for pls/pcr}
\usage{
predict_pls_pcr(model, newdata)
}
\arguments{
\item{model}{pls/pcr model}

\item{newdata}{new data}
}
\description{
predict method for pls/pcr
}
\keyword{internal}
