% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble-analysis.R
\name{prettify}
\alias{prettify}
\title{prettify}
\usage{
prettify(WL)
}
\arguments{
\item{WL}{Results from Significance experiments}
}
\value{
Pretty results from Significance experiments
}
\description{
Utility function that transforms the results from \code{WinLoss}
into a more visual appealing frame, to facilitate analysis.
}
