% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classmethods-hats.R
\docType{methods}
\name{prop_hat,tse_hat-method}
\alias{prop_hat,tse_hat-method}
\title{Proportionalize the predictions of base models}
\usage{
\S4method{prop_hat}{tse_hat}(object)
}
\arguments{
\item{object}{object of class \code{tse_hat}}
}
\description{
Proportionalize the predictions of base models
}
