% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classmethods-hats.R
\name{prop_hat}
\alias{prop_hat}
\title{Proportionalize the predictions of base models}
\usage{
prop_hat(object)
}
\arguments{
\item{object}{object of class \code{tse_hat}}
}
\description{
Proportionalize the predictions of base models
}

