% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proportion}
\alias{proportion}
\title{proportion}
\usage{
proportion(score, ...)
}
\arguments{
\item{score}{a numeric vector.}

\item{...}{Further arguments to \code{var} and \code{sum} function (e.g. na.rm = TRUE)}
}
\value{
A vector of proportions
}
\description{
Utility function used to compute the proportion of the values of a vector
The proportion of a value is its ratio relative
to the sum of the vector.
}
\examples{
proportion(rnorm(5L))
proportion(1:10)

}

