% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca-buildC.r
\name{prune_c_outperformance}
\alias{prune_c_outperformance}
\title{Prune subsets by out-performance}
\usage{
prune_c_outperformance(loss_info, lambda, depth = NULL)
}
\arguments{
\item{loss_info}{loss data in unseen observations}

\item{lambda}{smoothing window size}

\item{depth}{depth size how large is the maximum size
of the subsets. If NULL, defaults to no. of predictors
minus one.}
}
\description{
Prune subsets by out-performance
}
\keyword{internal}
