% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.r
\name{r_squared}
\alias{r_squared}
\title{Computing R squared}
\usage{
r_squared(y, y_hat)
}
\arguments{
\item{y}{A numeric vector representing the actual values.}

\item{y_hat}{A numeric vector representing the forecasted values.}
}
\description{
Also known as the coefficient of determination
}
\seealso{
Other error/performance functions: \code{\link{ae}},
  \code{\link{mae}}, \code{\link{mase}}, \code{\link{mse}},
  \code{\link{se}}
}
\keyword{internal}
