% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-models.R
\name{RandomForest}
\alias{RandomForest}
\title{learningRandomForest}
\usage{
RandomForest(form, data, learner.pars)
}
\arguments{
\item{form}{formula}

\item{data}{embedded time series}

\item{learner.pars}{list of learner parameter.}
}
\description{
learningRandomForest is the implementation of learning procedure
of an Random Forest model based on the implementation of \code{ranger} package.
}

