% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-preprocess.R
\name{rangedvar}
\alias{rangedvar}
\title{Normalized variance by range}
\usage{
rangedvar(x, ...)
}
\arguments{
\item{x}{time series as vector}

\item{...}{further params to \code{var}, \code{max} and \code{min}}
}
\description{
Normalized variance by range
}

