% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tse-pipes.r
\name{readSignal}
\alias{readSignal}
\title{readSignal}
\usage{
readSignal(raw_signal)
}
\arguments{
\item{raw_signal}{a new embedding vector to process into prediction-ready format}
}
\value{
A munged embedding vector
}
\description{
readSignal is an utility function used to munge a new data point into
the \code{data.frame} format identic to a given embedded time series.
Using this application turns a \code{raw_signal} into a format suitable for
tseModel objects to operate.
}

