% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembling-pipes.R
\name{recent_lambda_observations}
\alias{recent_lambda_observations}
\title{Get most recent lambda observations}
\usage{
recent_lambda_observations(data, lambda)
}
\arguments{
\item{data}{time series data as data.frame}

\item{lambda}{number of observations to keep}
}
\description{
Get most recent lambda observations
}
\keyword{internal}
