% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembler.R
\name{Regret}
\alias{Regret}
\title{Weighting base Models by Regret}
\usage{
Regret(SE, Mstar)
}
\arguments{
\item{SE}{Squared Error of the models at each test point}

\item{Mstar}{Maximum expected loss}
}
\description{
Weighting base Models by Regret
}

