% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roll_mean_matrix}
\alias{roll_mean_matrix}
\title{Computing the rolling mean of the columns of a matrix}
\usage{
roll_mean_matrix(x, lambda)
}
\arguments{
\item{x}{a numeric data.frame;}

\item{lambda}{periods to average over when computing the
moving average.}
}
\description{
Computing the rolling mean of the columns of a matrix
}
\keyword{internal}
