% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rollmeanmatrix}
\alias{rollmeanmatrix}
\title{Auxiliary function used to compute column-wise moving averages on a matrix}
\usage{
rollmeanmatrix(mat, ma.N)
}
\arguments{
\item{mat}{matrix}

\item{ma.N}{periods to average over when computing the moving average.}
}
\description{
Auxiliary function used to compute column-wise moving averages on a matrix
}
