% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflows-baselines.r
\name{s_champ}
\alias{s_champ}
\title{ss_champ}
\usage{
s_champ(form, train, test, learner, learner.pars, varying.embed = TRUE,
  varying.trainwindow = TRUE, ...)
}
\arguments{
\item{form}{form}

\item{train}{train}

\item{test}{test}

\item{learner}{learner}

\item{learner.pars}{learner pars}

\item{varying.embed}{ve}

\item{varying.trainwindow}{vw}

\item{...}{adsad}
}
\description{
ss_champ
}

