% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflows-baselines.r
\name{s_w}
\alias{s_w}
\title{ss_W}
\usage{
s_w(form, train, test, learner, learner.pars, varying.embed = FALSE,
  varying.trainwindow = FALSE, ...)
}
\arguments{
\item{form}{form}

\item{train}{train}

\item{test}{test}

\item{learner}{learner}

\item{learner.pars}{learner pars}

\item{varying.embed}{ve}

\item{varying.trainwindow}{vw}

\item{...}{adsa}
}
\description{
ss_W
}

