% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflows-baselines.r
\name{S_W50}
\alias{S_W50}
\title{Forecaster Combination sw50}
\usage{
S_W50(form, train, test, learner, learner.pars, ma_n = 50, ...)
}
\arguments{
\item{form}{Formula}

\item{train}{embedded time series used for training the base learners}

\item{test}{embedded time series used for testing}

\item{learner}{Character vector describing the base algorithms to be trained.}

\item{learner.pars}{Named list describing the parameter of the \code{learner}. Below are
described some examples.}

\item{ma_n}{moving avg}

\item{...}{Further parameters to pass to the function}
}
\description{
Forecaster Combination sw50
}
