% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-models.R
\name{SAE}
\alias{SAE}
\title{SAE}
\usage{
SAE(form, data, learner.pars)
}
\arguments{
\item{form}{formula}

\item{data}{embedded time series}

\item{learner.pars}{list of learner parameter.}

\item{...}{Another parameters to pass to deepnet::sae.dnn.train}
}
\description{
SAE is the implementation of learning procedure
of a Stacked Auto Encoder model based on the
implementation of \code{deepnet} package.
We advise not to use this model unless you are
an expert in neural networks. Tunable parameters are
Activation function, hidden units, hidden units dropout
and visible units dropout.
}
