% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembling-pipes.R
\name{select_best}
\alias{select_best}
\title{Selecting best model according to weights}
\usage{
select_best(model_scores)
}
\arguments{
\item{model_scores}{matrix containing the model weights
across the observations}
}
\description{
This function select the best model from a matrix of data x models.
For each row (data point), the model with maximum weight is assigned
a weight of 1, while the remaining models are assigned a weight of 0.
}
\seealso{
Other weighting base models: \code{\link{EMASE}},
  \code{\link{build_committee}},
  \code{\link{get_top_models}},
  \code{\link{model_recent_performance}},
  \code{\link{model_weighting}}
}
\concept{weighting base models}
\keyword{internal}
