% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential-reweight.R
\name{sequential_reweighting}
\alias{sequential_reweighting}
\title{Sequential Re-weighting for controlling predictions' redundancy}
\usage{
sequential_reweighting(sliding_similarity, W)
}
\arguments{
\item{sliding_similarity}{list of pairwise similarity values. See
\code{\link{sliding_similarity}}}

\item{W}{weights before re-weighting}
}
\description{
Besides ensemble heterogeneity we encourage diversity
explicitly during the aggregation of the output of experts.
This is achieved by taking into account not only predictions
of performance produced by the arbiters, but also the
correlation among experts in a recent window of observations.
}
\keyword{internal}
