% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta-learn.r
\name{setup.metadata}
\alias{setup.metadata}
\title{setup metadata using oob samples and ae of models}
\usage{
setup.metadata(OOB.train, test, Y_hat, dynamics.FUN = NULL)
}
\arguments{
\item{OOB.train}{oob train and ae info returned by \code{ForwardValidation} fun,
combined with \code{OOB.fun}.}

\item{test}{test set}

\item{Y_hat}{M prediction for the test set}

\item{dynamics.FUN}{function that calculates dynamics}
}
\value{
metadata ready for metalearning the overlineM
}
\description{
using forward validation we get the oob train data and
oob ae of models to augment our meta set
here we combine these to the test info to setup the meta data
to train the overlineM
}
