% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential-reweight.R
\name{sliding_similarity}
\alias{sliding_similarity}
\title{Sliding similarity via Pearson's correlation}
\usage{
sliding_similarity(Y_hat_ext, lambda)
}
\arguments{
\item{Y_hat_ext}{Predictions from the base-learners
across the examples.}

\item{lambda}{window size for computing correlations}
}
\value{
a list with a correlation matrix
for each prediction point
}
\description{
Sliding similarity via Pearson's correlation
}
