% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{softmax}
\alias{softmax}
\title{Computing the softmax}
\usage{
softmax(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
This function computes the softmax function in
a numeric vector
}
\keyword{internal}
