% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_by}
\alias{split_by}
\alias{split_by_}
\alias{split_by.}
\title{Splitting expressions by pattern}
\usage{
split_by(expr, split, unlist. = TRUE, ...)

split_by_(expr, ...)

split_by.(expr, ...)
}
\arguments{
\item{expr}{character expression to split;}

\item{split}{expression to split \code{expr} by;}

\item{unlist.}{Logical. If TRUE, the splitted \code{expr} is unlisted;}

\item{...}{Further parameters to pass to \code{strsplit};}
}
\value{
a list or vector with a splitted expression
}
\description{
This is an utility function that can be used to split expressions.
It is based on \code{strsplit} function.
split_by is the general purpose \emph{splitter}
split_by_ splits expressions by \"_\"
split_by. splits expressions by a dot
}
\examples{
split_by_("time_series")
split_by.("time.series")
split_by("born2bewild", "2")

}
\keyword{internal}
