% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_vec}
\alias{split_vec}
\title{split_vec}
\usage{
split_vec(x, n, avg = TRUE)
}
\arguments{
\item{x}{vector to split}

\item{n}{number of parts to split x into}

\item{avg}{Logical. If \code{TRUE} returns the results averaged by mean.}
}
\value{
List of splitted vectors
}
\description{
Utility function that splits a vector into n parts
In this package this function is used for smoothing large vectors for plotting
}
\examples{
split_vec(letters[1:6], 3, F)

}
