% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-models.R
\name{SVM}
\alias{SVM}
\title{SVM}
\usage{
SVM(form, data, learner.pars = NULL)
}
\arguments{
\item{form}{formula}

\item{data}{embedded time series}

\item{learner.pars}{list of learner parameter.}
}
\description{
SVM is the implementation of learning procedure
of an SVM for regression based on the implementation of \code{kernlab} package.
}

