% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-models.r
\name{tbats_predict}
\alias{tbats_predict}
\title{Predict function for a model created from bm_tbats}
\usage{
tbats_predict(model, test)
}
\arguments{
\item{model}{model created using \strong{bm_tbats} function}

\item{test}{test data created using \strong{embed_timeseries} function}
}
\description{
Predict function for a model created from bm_tbats
}
\keyword{internal}
