% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-preprocess.R
\name{ts.dynamics.complete}
\alias{ts.dynamics.complete}
\title{Time series dynamics}
\usage{
ts.dynamics.complete(data)
}
\arguments{
\item{data}{Numeric matrix to compute dynamics in, such as an
embedded time series.}
}
\description{
Time series dynamics
}

