% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflows-arima.r
\name{tse.arch11arma}
\alias{tse.arch11arma}
\title{ARCH(1,1)-ARMA process workflow}
\usage{
tse.arch11arma(form, train, test, modelOutput = FALSE, ...)
}
\arguments{
\item{form}{Formula}

\item{train}{embedded time series used for training the base learners}

\item{test}{embedded time series used for testing}

\item{modelOutput}{Logical. If TRUE, the methods applied to \code{tsensembler-class} return
a predictive model ensemble. Otherwise, it returns performance statistics based on
\code{\link[performanceEstimation]{performanceEstimation}}.}

\item{...}{Further parameters to \code{arch11armaWorkflow}}
}
\description{
The forecast is calculated in a point-wise fashion. That is, the process
is estimated at each step.
}

