% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflows-arima.r
\name{tse.arima}
\alias{tse.arima}
\title{arimaWorkflow}
\usage{
tse.arima(form, train, test, modelOutput = FALSE, ...)
}
\arguments{
\item{form}{Formula, e.g. \code{target ~.}}

\item{train}{embedded time series used for training the base learners}

\item{test}{embedded time series used for testing}

\item{modelOutput}{Logical. If TRUE, the methods applied to \code{tsensembler-class} return
a predictive model ensemble. Otherwise, it returns performance statistics based on
\code{\link[performanceEstimation]{performanceEstimation}}.}

\item{...}{Further parameters to pass to \code{tse.arima}}
}
\value{
a list containing the true values and predicted values for test set
}
\description{
This function implements a workflow of autoregressive and moving average models.
It is based on \code{auto.arima} function of \code{forecast} package, which
performs an optimal parameter tuning.
}
\seealso{
\code{\link[forecast]{auto.arima}} for an automatic arima procedure.
}
