% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflows-arima.r
\name{tse.arima_xreg}
\alias{tse.arima_xreg}
\title{auto arima with external features}
\usage{
tse.arima_xreg(form, train, test, xreg.id, ...)
}
\arguments{
\item{form}{formula}

\item{train}{train set}

\item{test}{test set}

\item{xreg.id}{id of external feats}

\item{...}{Further parameters to the workflow}
}
\description{
auto arima with external features
}
