% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdef-hats.r
\docType{class}
\name{tse_hat-class}
\alias{tse_hat-class}
\title{Predictions produced by a \code{tseModel} object}
\value{
Set of results from predicting with a \code{tseModel}
ensemble
}
\description{
Predictions produced by a \code{tseModel} object
}
\section{Slots}{

\describe{
\item{\code{y_hat}}{Combined predictions. Numeric vector}

\item{\code{Y_hat}}{List of predictions made by individual models}

\item{\code{Y_W}}{Normalized weights of individual models}

\item{\code{Y_committee}}{List of top models at each prediction step}

\item{\code{Y_rank}}{List of rank of models at each prediction step}

\item{\code{data}}{data used to produce predictions}
}}

\seealso{
\code{\link{tseModel-class}} for the \code{tseModel} class.
}
