% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdef-hats.r
\name{tse_hat}
\alias{tse_hat}
\title{Predictions by a tseModel}
\usage{
tse_hat(y_hat = "numeric", Y_hat = "OptionalList", Y_W = "OptionalDF",
  Y_committee = "OptionalList", Y_rank = "OptionalList",
  data = "data.frame")
}
\arguments{
\item{y_hat}{Combined predictions. Numeric vector}

\item{Y_hat}{List of predictions made by individual models}

\item{Y_W}{Normalized weights of individual models}

\item{Y_committee}{List of top models at each prediction step}

\item{Y_rank}{List of rank of models at each prediction step}

\item{data}{data used to produce predictions}
}
\value{
Set of results from predicting with a \code{tseModel}
ensemble
}
\description{
Predictions by a tseModel
}
\seealso{
\code{\link{tseModel-class}} for the \code{tseModel} class.
}

