% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdef-tsesearch.R
\docType{class}
\name{tsesearch-class}
\alias{tsesearch-class}
\title{tsesearch}
\description{
An S4 class that combines the meta-data about the Time Series Ensemble.
It wraps the ensemble procedure on top of the performance estimation workflows.
}
\section{Slots}{

\describe{
\item{\code{timeseries}}{A time series or list of time series of class \code{tsSource}.
This class inherits the classes \code{xts}, \code{zoo} and \code{list}.}

\item{\code{workflow}}{Object of class Workflow. Check \code{\link[performanceEstimation]{Workflow-class}} for
further details.}

\item{\code{embedding.dimension}}{The maximum embedding dimension used to transform the series.
If \code{timeseries} is a single time series and \code{embedding.dimension} is \emph{NULL}
the parameter will be estimated using function.}

\item{\code{nReps}}{Number of Monte Carlo repetitions to perform the experiment simulations.
Defaults to 10. More information in \code{\link[performanceEstimation]{MonteCarlo-class}}.}

\item{\code{szTrain}}{Ratio of the available window size used for training.
More information at \code{\link[performanceEstimation]{MonteCarlo-class}}}

\item{\code{szTest}}{Ratio of the available window size used for testing.
More information at \code{\link[performanceEstimation]{MonteCarlo-class}}}

\item{\code{modelOutput}}{Logical. If TRUE, the methods applied to \code{tsensembler-class} return
a predictive model ensemble. Otherwise, it returns performance statistics based on
\code{\link[performanceEstimation]{performanceEstimation}}.}
}}
\examples{
\dontrun{
TSEr <- tsensembler(timeseries = ts,
                    workflow = Workflow(wf = 'TSE',
                                        learner = c('MARS', 'PPR'),
                                        learner.pars = list(mars = list(nk = c(5, 2),
                                                                        degree= c(3, 4)),
                                                            ppr = list(nterms = c(2,3,4))),
                                        varying.embed = TRUE,
                                        varying.trainwindow = FALSE,
                                        committee.ratio = .1,
                                        combine.type = "regret_theorem23",
                                        verbose = FALSE),
                    modelOutput = TRUE,
                    embedding.dimension = 30,
                    nReps = 2)
}

}
\seealso{
\code{\link[performanceEstimation]{performanceEstimation}} for the ensemble estimation
procedure and \code{\link[performanceEstimation]{MonteCarlo-class}} for Monte Carlo simulation setup.
The methods \code{\link{ensembler}} or \code{\link{Multi.ensembler}} are used to run an
object of class \code{tsensembler}. The normal output of the previously mentioned methods are
of class \code{\link{tseModel-class}}.
}

