% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classdef-tseSpecs.r
\name{tseSpecs}
\alias{tseSpecs}
\title{Setup base learning models}
\usage{
tseSpecs(learner = "character", learner_pars = "OptionalList",
  k_size = "OptionalNumeric", varying_k = "OptionalNumeric",
  varying_w = "OptionalNumeric", data = "OptionalDF")
}
\arguments{
\item{learner}{Character vector with the base learners
to be trained (e.g. \strong{MARS})}

\item{learner_pars}{List with parameter setting for the
\code{learner}}

\item{k_size}{Embedding dimension use to embed the time series
into an Euclidean space}

\item{varying_k}{An embed split parameter used to encourage
data diversity into the ensemble. Each \code{learner} is trained
in each setting of \strong{varying_k}. For example, a
\strong{varying_k} equal to the \code{k_size} uses the whole
embedding dimension; A \strong{varying_k} as \code{c(k_size,
k_size / 2)} trains the base models in the whole set of embed
cols and in the set of the \strong{varying_k / 2} most
recent embeds. Defaults to the whole set (\strong{k_size})}

\item{varying_w}{A data split parameter used to encourage
data diversity into the ensemble. Similar to \code{varying_k}
parameter but related to the embedding vectors instead of the
embedding colums. Defaults to the whole data set \code{nrow(data)}}

\item{data}{embedded time series of class \code{data.frame}}
}
\description{
This class sets up the base learning models and respective
parameters setting to learn the ensemble.
}
