% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-preprocess.R
\name{unembed.timeseries}
\alias{unembed.timeseries}
\title{unembed.timeseries}
\usage{
unembed.timeseries(train, test = NULL)
}
\arguments{
\item{train}{train series data.frame}

\item{test}{test series data.frame}
}
\value{
a time series
}
\description{
This function reverts the process of embed.timeseries function.
It takes an embedded series and returns the series as a sequential vector.
}

