% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classmethods-tsensembler.R
\name{univariate.ensembler}
\alias{univariate.ensembler}
\title{Learning a Dynamic Ensemble for Time Series Forecasting}
\usage{
univariate.ensembler(obj)
}
\arguments{
\item{obj}{An object of class \code{tsensembler}.}
}
\value{
A predictive model ensemble of class \code{tseModel}.
}
\description{
This is the implementation of the learning procedure of the ensemble method.
The application is run on an object of class \code{tsensembler}, which provide
the options about how the ensemble should be construct such as: the embedding
dimension; the number of Monte Carlo repetitions, etc.
This method differs from \code{gs.ensembler} in the fact that \emph{ensembler}
does not estimate several models. \emph{ensembler} simply trains an Dynamic
Ensemble according to the specs in the object \strong{obj} of class \code{tsensembler}.
}
\examples{

\dontrun{
require(xts)
ts <- as.xts(rnorm(100L), order.by = Sys.Date() + rnorm(100L))
TSEr <- tsensembler(timeseries = ts,
                    form = target ~.,
                    workflow = Workflow(wf = 'TSE',
                                        learner = c('MARS', 'PPR'),
                                        learner.pars = list(mars = list(nk = c(5, 2),
                                                                        degree= c(3, 4)),
                                                            ppr = list(nterms = c(2,3,4))),
                                        varying.embed = TRUE,
                                        varying.trainwindow = FALSE,
                                        committee.ratio = .1,
                                        aggregationFUN = "emase-scommittee",
                                        verbose = FALSE),
                    embedding.dimension = 30,
                    nReps = 2,
                    modelOutput = TRUE)
tse.model <- ensembler(TSEr)
}

}
\seealso{
\code{\link{tsensembler-class}} for details about the \strong{obj} and
\code{\link{tseModel-class}} for details on the generated model.
\code{\link[performanceEstimation]{MonteCarlo-class}} comprises an explanation
about the Monte Carlo simulation procedure. \code{\link{TSE}} for details on the workflow
for the time series ensemble.
}

