% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vlapply}
\alias{vlapply}
\title{vapply extension for logical values}
\usage{
vlapply(x, fun, ..., use.names = FALSE)
}
\arguments{
\item{x}{A vector (atomic or list)}

\item{fun}{function to be applied}

\item{...}{optional arguments to fun}

\item{use.names}{logical. Check argument USE.NAMES from \code{vapply} function.}
}
\description{
vapply extension for logical values
}
\seealso{
\code{\link{vapply}}
}

