% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble-analysis.R
\name{WinLoss}
\alias{WinLoss}
\title{Statistical Wins and Losses}
\usage{
WinLoss(pairedComp, metric = "mse", test = "t.test", pretty = FALSE,
  pvalue = 0.05)
}
\arguments{
\item{pairedComp}{Paired Comparison results. This is a list
with experiments computed from the mean and median scores of
workflows.}

\item{metric}{Metric used to compare methods. Defaults to Mean Squared Error (MSE).}

\item{test}{Statistical significance test to perform. Defaults to \code{WilcoxonSignedRank.test}.
The other option is the \code{t.test}}

\item{pretty}{Logical: If TRUE, results are rendered in a more visually appealling way.}

\item{pvalue}{Significance threshold}
}
\value{
a data.frame with the significance test results accross several experiments
}
\description{
WinLoss is an utility function built of top of \code{pairedComparisons}
method (\code{performanceEstimation} package) that analyses the
performance of models. Particularly, \code{WinLoss} probes for statistically
significant differences in the performance of models.
}
\examples{

\dontrun{
pairedComps <- pairedComparisons(ComparisonResults.obj, baseline)
wl <- WinLoss(pairedComps, "WilcoxonSignedRank.test")
}

}
\seealso{
\code{\link[performanceEstimation]{pairedComparisons}}
}

