% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model-xgb.r
\name{xgb_optimizer}
\alias{xgb_optimizer}
\title{XGB optimizer}
\usage{
xgb_optimizer(X, y, gsearch)
}
\arguments{
\item{X}{Covariates}

\item{y}{Target values}

\item{gsearch}{Grid search}
}
\description{
XGB optimizer
}
