% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model-xgb.r
\name{xgb_predict}
\alias{xgb_predict}
\title{XGBoost predict function}
\usage{
xgb_predict(model, newdata)
}
\arguments{
\item{model}{Model from bm_xgb}

\item{newdata}{Test data}
}
\description{
XGBoost predict function
}
\keyword{internal}
