% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model-xgb.r
\name{bm_xgb}
\alias{bm_xgb}
\title{xgb base model}
\usage{
bm_xgb(form, data, lpars)
}
\arguments{
\item{form}{formula}

\item{data}{training data}

\item{lpars}{learning parameters}
}
\description{
xgb base model
}
\keyword{internal}
