% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca-integration.r
\name{combine_committees}
\alias{combine_committees}
\title{Merge across sub-ensembles}
\usage{
combine_committees(predict_info, FUN, object, newdata)
}
\arguments{
\item{predict_info}{data from ...}

\item{FUN}{function used to combine the models in each
subset.}

\item{object}{constructive aggregation class object}

\item{newdata}{new data to make predictions}
}
\description{
Merge across sub-ensembles
}
\keyword{internal}
