% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembler.R
\name{EMASE}
\alias{EMASE}
\title{Weighting Base Models by their Moving Average Squared Error}
\usage{
EMASE(SE, ma.N, preweights, smartass = FALSE)
}
\arguments{
\item{SE}{Squared Error of the models at each test point}

\item{ma.N}{Number of periods to average over when computing MASE}

\item{preweights}{pre-weights of the base models computed in the
train set.}

\item{smartass}{if TRUE, lastpoint is used to compute EMASE, 
breaking the golden rule. Defaults to FALSE. Useful for hindsight analysis.}
}
\value{
The weights of the models in test time.
}
\description{
This function computes the weights of the learning models
using the Moving Average Squared Error \(MASE\) function
This method provides a simple way to quantify the recent
performance of each base learner and adapt the combined
model accordingly.
}

