% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{erfc}
\alias{erfc}
\title{Erfc}
\usage{
erfc(x, alpha = 2)
}
\arguments{
\item{x}{A numeric vector in the range 0-1}

\item{alpha}{parameter used to sharp the Erfc curve. Defaults to 1.}
}
\value{
The complementary Gaussian error
}
\description{
Erfc stands for the Complementary Gaussian Error Function.
This mathematical formula can be used as a squashing function.
}
\examples{

\dontrun{
  erfc(.1)
  erfc(c(.1, .7))
}

}
