% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tse-pipes.r
\name{get_y}
\alias{get_y}
\title{Get true values from test set
from formula and the test set}
\usage{
get_y(test, form)
}
\arguments{
\item{test}{test set}

\item{form}{formula}
}
\description{
Get true values from test set
from formula and the test set
}

