% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out-of-bag-pipes.r
\name{holdout}
\alias{holdout}
\title{Holdout}
\usage{
holdout(x, beta, FUN, ...)
}
\arguments{
\item{x}{data to split into \code{nfolds} blocks;}

\item{beta}{ratio of observations for training}

\item{FUN}{function to apply to train/test split}

\item{...}{further arguments to FUN}
}
\description{
Holdout
}
\keyword{internal}
