% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out-of-bag-pipes.r
\name{intraining_estimations}
\alias{intraining_estimations}
\title{Out-of-bag loss estimations}
\usage{
intraining_estimations(train, test, form, specs)
}
\arguments{
\item{train}{train set from the training set;}

\item{test}{test set from the training set;}

\item{form}{formula;}

\item{specs}{object of class \code{\link{model_specs-class}}. Contains
the specifications of the base models.}
}
\value{
A list containing two objects:
\describe{
\item{mloss}{loss of base models in \strong{test}}
\item{oob}{out-of-bag test samples}
}
}
\description{
A pipeline for retrieving out-of-bag loss estimations
}
\seealso{
Other out-of-bag functions: \code{\link{intraining_predictions}}
}
