% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{l1apply}
\alias{l1apply}
\title{lapply on the rows}
\usage{
l1apply(obj, FUN, ...)
}
\arguments{
\item{obj}{obj}

\item{FUN}{func}

\item{...}{Further parameters to \code{lapply}}
}
\description{
lapply on the rows
}

