% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembler.R
\name{select_best}
\alias{select_best}
\title{Select best model from weights}
\usage{
select_best(model_scores)
}
\arguments{
\item{model_scores}{model weights}
}
\description{
Select best model from weights
}
