% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembler.R
\name{select_best}
\alias{select_best}
\title{Selecting best model according to weights}
\usage{
select_best(model_scores)
}
\arguments{
\item{model_scores}{Matrix containing the model weights across the data}
}
\description{
Selecting best model according to weights
}
