% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arbitration-methods.r
\name{train_ade}
\alias{train_ade}
\title{Training procedure of for ADE}
\usage{
train_ade(form, train, specs, lambda, lfun, meta_model_type)
}
\arguments{
\item{form}{formula;}

\item{train}{training data as a data frame;}

\item{specs}{a \code{\link{model_specs-class}} object class. It contains
the parameter setting specifications for training the ensemble;}

\item{lambda}{window size. Number of observations to compute
the recent performance of the base models, according to the
committee ratio \strong{omega}. Essentially, the top \emph{omega}
models are selected and weighted at each prediction instance, according
to their performance in the last \emph{lambda} observations.
Defaults to 50 according to empirical experiments;}

\item{lfun}{meta loss function}
}
\description{
Base level models are trained according to \strong{specs},
and meta level models are trained using a blocked prequential
procedure in out-of-bag samples from the training data.
}
\keyword{internal}
