% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-model-xgb.r
\name{xgb_predict}
\alias{xgb_predict}
\title{xgb predict fun}
\usage{
xgb_predict(model, newdata)
}
\arguments{
\item{model}{model from bm_xgb}

\item{newdata}{newdata}
}
\description{
xgb predict fun
}
\keyword{internal}
